# Store Object signals in a HashMap rather than a VMap