Fixes #69874