@akien-mga 2023-01-30
Can you share benchmark results to assess the gain?