@myaaaaaaaaa 2023-01-30
Sure, with the caveat that these numbers are from a debug build

Code:
```
extends Benchmark

func build_node(num_signals: int) -> Node3D:
	var mesh := ArrayMesh.new()
	var rt := MeshInstance3D.new()
	for i in num_signals:
		var n := MeshInstance3D.new()
		rt.add_child(n)

		# Implicitly connects a changed signal
		n.mesh = mesh

	return rt

func benchmark_signals():
	for i in [2,20,200,2000]:
		print()
		print("num_signals: ", i)
		var m := Time.get_ticks_usec()
		for j in 10: #More reliable timing
			build_node(i).free()
		print("%.3f ms" % ((Time.get_ticks_usec() - m) / 10 / 1e3))
```

Before: ( Using #72346 for consistent behavior, see that for details )
```
num_signals: 2
0.223 ms

num_signals: 20
1.612 ms

num_signals: 200
25.631 ms

num_signals: 2000
1421.480 ms
```

After: ( Note that scaling is now linear )
```
num_signals: 2
0.221 ms

num_signals: 20
1.368 ms

num_signals: 200
13.461 ms

num_signals: 2000
124.343 ms
```