@reduz 2023-04-07
The only problem with this approach is that emission could be slower. It would be nice to benchmark this in exchange.