@akien-mga 2023-04-25
> The only problem with this approach is that emission (which is more critical than connection time) could be slower. It would be nice to benchmark this in exchange.

@myaaaaaaaaa Would you have time to do a quick test for this?